+++ src/components/common/modal/ButtonSelectModal.js
@@ -8,7 +8,7 @@ import {
   submitForm,
   postWriteSelector,
   setField,
-} from '../../../modules/tommunity/postWrite';
+} from '../../../modules/community/postWrite';
 import { SelectButton } from '../../community/PostWrite/SelectButton';
 import { SelectAllButton } from '../../community/PostWrite/SelectAllButton';
 import { List, Border } from '../../../styles/ModalStyle';
