+++ src/components/community/comment/commentOutput.js
@@ -1,7 +1,7 @@
 /**CommentOutPut 댓글을 출력하는 컴포넌트 */
 import { useDispatch } from 'react-redux';
 import { CommentBox } from '../../../styles/community/CommentStyle';
-import { changeCommentField } from '../../../modules/tommunity/postDetail';
+import { changeCommentField } from '../../../modules/community/postDetail';
 import ReplyOutPut from './replyOutput';
 import useModal from '../../common/modal/useModal';
 import { Profile } from './Profile';
