+++ src/containers/community/PostList.js
@@ -1,7 +1,7 @@
 /**PostList 커뮤니티 게시물 리스트가 나오는 컨테이너 */
 import { useSelector, useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { changePostDetailField } from '../../modules/tommunity/postDetail';
+import { changePostDetailField } from '../../modules/community/postDetail';
 import PostItem from '../../components/community/PostList/PostItem';
 import useListQuery from '../../modules/queries/useListQuery';
 import { ErrorMessage } from './ErrorMessage';
