+++ src/index.js
@@ -15,9 +15,9 @@ import userSlice from './modules/user';
 import activitySlice from './modules/activity';
 import barSlice from './modules/bar';
 import filterSlice from './modules/filter.js';
-import postWriteSlice from './modules/tommunity/postWrite.js';
-import PostDetailSlice from './modules/tommunity/postDetail.js';
-import { modalSlice } from './modules/tommunity/modal.js';
+import postWriteSlice from './modules/community/postWrite.js';
+import PostDetailSlice from './modules/community/postDetail.js';
+import { modalSlice } from './modules/community/modal.js';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import './styles/common.scss';
