+++ src/modules/community/postDetail.js
